moved data set input validation to schema not model
