+++ station/app/models/datasets.py
@@ -9,37 +9,16 @@ from datetime import datetime
 from station.app.db.base_class import Base
 
 
-class StorageType(Enum):
-    """
-    Enum for storage types
-    """
-    LOCAL = "local"
-    MINIO = "minio"
-    DB = "db"
-
-
-class DataType(Enum):
-    """
-    Enum for data types
-    """
-    IMAGE = "image"
-    GENOME = "genome"
-    FHIR = "fhir"
-    CSV = "csv"
-    STRUCTURED = "structured"
-    UNSTRUCTURED = "unstructured"
-    HYBRID = "hybrid"
-
 
 class DataSet(Base):
     __tablename__ = "datasets"
     id = Column(Integer, primary_key=True, index=True)
     created_at = Column(DateTime, default=datetime.now())
     updated_at = Column(DateTime, nullable=True)
-    proposal_id = Column(Integer, nullable=True)
+    proposal_id = Column(String, nullable=True)
     name = Column(String)
-    data_type = Column(String, default=DataType.FHIR.value)
-    storage_type = Column(String, default=StorageType.MINIO.value)
+    data_type = Column(String, nullable=True)
+    storage_type = Column(String, nullable=True)
     access_path = Column(String, nullable=True)
     fhir_server = Column(Integer, ForeignKey('fhir_servers.id'), nullable=True)
     summary = Column(JSON, nullable=True)
